# standard/backend/docs/swagger.yaml
basePath: /api/v1/standard
definitions:
...
  github_com_addp_standard_internal_models.Element:
    properties:
      classification_id:
        description: 关联数据分类
        type: integer
      code:
        type: string
      code_set_id:
        description: 关联码值集
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      data_type:
        description: string/int/float/date/bool/json
        type: string
      default_value:
        type: string
      definition:
        type: string
      domain_id:
        type: integer
      example_values:
        items:
          type: string
        type: array
      format:
        type: string
      id:
        type: integer
      length:
        type: integer
      name:
        type: string
      nullable:
        type: boolean
      precision_num:
        type: integer
      quality_rules:
        allOf:
        - $ref: '#/definitions/github_com_addp_standard_internal_models.JSONB'
        description: 质量规则
      scale:
        type: integer
      security_level:
        description: L1/L2/L3/L4
        type: string
      status:
        description: draft/approved/deprecated
        type: string
      steward_id:
        type: integer
      tags:
        items:
          type: string
        type: array
      tenant_id:
        type: integer
      unit_id:
        description: 关联计量单位
        type: integer
      updated_at:
        type: string
      updated_by:
        type: integer
      value_range:
        $ref: '#/definitions/github_com_addp_standard_internal_models.JSONB'
      version:
        type: integer
    type: object
...
  github_com_addp_standard_internal_models.JSONB:
    additionalProperties: true
    type: object
...
  github_com_addp_standard_internal_models.PaginatedElementResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/github_com_addp_standard_internal_models.Element'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_addp_standard_internal_models.UpdateElementRequest:
    properties:
      classification_id:
        type: integer
      code_set_id:
        type: integer
      data_type:
        type: string
      default_value:
        type: string
      definition:
        type: string
      domain_id:
        type: integer
      example_values:
        items:
          type: string
        type: array
      format:
        type: string
      length:
        type: integer
      name:
        type: string
      nullable:
        type: boolean
      precision_num:
        type: integer
      quality_rules:
        additionalProperties: true
        type: object
      scale:
        type: integer
      security_level:
        type: string
      steward_id:
        type: integer
      tags:
        items:
          type: string
        type: array
      unit_id:
        type: integer
      value_range:
        additionalProperties: true
        type: object
      version:
        type: integer
    required:
    - version
    type: object
...
  github_com_addp_standard_internal_models.VersionRequest:
    properties:
      version:
        type: integer
    required:
    - version
    type: object
...
host: localhost:8110
info:
  contact: {}
  title: ADDP Standard API
  version: "1.0"
paths:
...
  /elements:
    get:
      parameters:
      - description: 数据元 ID 集合，逗号分隔，最多 100 个 | Data element IDs, comma-separated,
          maximum 100
        in: query
        name: ids
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_addp_standard_internal_models.PaginatedElementResponse'
        "400":
          description: 无效的 ID 集合 | Invalid ID collection
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取数据元列表 | List data elements
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.element.read
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建数据元 | Create data element
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.element.create
  /elements/{id}:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 删除数据元 | Delete data element
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.element.delete
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取数据元详情 | Get data element detail
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.element.read
    put:
      parameters:
      - description: 更新数据元 | Update data element
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_addp_standard_internal_models.UpdateElementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 资源版本冲突 | Resource version conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新数据元 | Update data element
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.element.update
  /elements/{id}/approve:
    post:
      parameters:
      - description: 当前资源版本 | Current resource version
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_addp_standard_internal_models.VersionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 资源版本冲突 | Resource version conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 审批数据元 | Approve data element
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.element.approve
...
  /elements/{id}/quality-rules:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录 | Authentication required
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: 无权访问 | Access denied
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取数据元质量规则 | Get data element quality rules
      tags:
      - Standard
      x-addp-auth-mode: permission
      x-addp-required-permissions:
      - standard.element.read
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"